refactor(elements): consent manager provider structure
Restructured the consent manager provider by:
- Splitting large file into multiple focused files
- Extracting translation and theme utility functions
- Improving type definitions and code organization
- Adding support for custom translations with more flexibility
