+++ packages/elements/src/react/common/store/use-consent-manager.ts
@@ -4,7 +4,7 @@ import type {
 	createConsentManagerStore,
 } from '@koroflow/core-js';
 import { useContext } from 'react';
-import { ConsentStateContext } from './consent-manager';
+import { ConsentStateContext } from '../context/consent-manager-context';
 
 /**
  * A custom React hook that provides access to the privacy consent state and management methods.
