+++ packages/elements/src/react/common/types/consent-manager.ts
@@ -1,0 +1,71 @@
+import type {
+	AllConsentNames,
+	ComplianceRegion,
+	ComplianceSettings,
+	NamespaceProps,
+	PrivacyConsentState,
+	TranslationConfig,
+	createConsentManagerStore,
+} from '@koroflow/core-js';
+import type { ReactNode } from 'react';
+
+/**
+ * Configuration options for the ConsentManagerProvider component.
+ *
+ * @remarks
+ * These props allow you to configure the initial state and behavior of the consent manager,
+ * including GDPR types, compliance settings, and namespace configuration.
+ *
+ * @public
+ */
+export interface ConsentManagerProviderProps extends NamespaceProps {
+	/**
+	 * @remarks
+	 * React elements to be rendered within the consent manager context.
+	 */
+	children: ReactNode;
+
+	/**
+	 * @remarks
+	 * Array of consent types to be pre-configured for GDPR compliance.
+	 * These types define what kinds of cookies and tracking are initially allowed.
+	 */
+	initialGdprTypes?: AllConsentNames[];
+
+	/**
+	 * @remarks
+	 * Region-specific compliance settings that define how the consent manager
+	 * should behave in different geographical regions.
+	 */
+	initialComplianceSettings?: Record<ComplianceRegion, ComplianceSettings>;
+
+	/**
+	 * @remarks
+	 * Whether to skip injecting default styles
+	 * @default false
+	 */
+	noStyle?: boolean;
+
+	/**
+	 * @remarks
+	 * Configuration for translations including available languages and settings.
+	 * Extends or overrides the default translations.
+	 */
+	translationConfig?: Partial<TranslationConfig>;
+}
+
+/**
+ * Internal context value interface for the consent manager.
+ *
+ * @remarks
+ * Combines both the current state and store instance for complete
+ * consent management functionality.
+ *
+ * @internal
+ */
+export interface ConsentManagerContextValue {
+	/** Current privacy consent state */
+	state: PrivacyConsentState;
+	/** Store instance for managing consent state */
+	store: ReturnType<typeof createConsentManagerStore>;
+}
