+++ packages/elements/src/react/theme/context.tsx
@@ -7,7 +7,7 @@
  */
 
 import { createContext, useContext } from 'react';
-import type { useConsentManager } from '../common/store/consent-manager';
+import type { useConsentManager } from '../common/store/use-consent-manager';
 
 /**
  * Configuration value type for the ThemeContext.
